Allow the user to be jumped to a specific anchor
This means there's not that annoying scrolling when you're half way down
the form and click on a button. Done by doing some processing of the URL
parameters in a process_form method, which then redirects through to the
main method with the appropriate #foo hash for taking the user to a
point, depending on which button was used to submit the form.