Fix issue when updating values in ANI calculator
The calculator values were being ignored when the adjusted_net_income
field was filled in.  This meant that if you updated any values in the
calculator, it wasn't updating the adjusted_net_income value unless you
first cleared it.

This changes it so that it will always calculate the adjusted_net_income
if any of the fields in the calculator form are filled in.