Update poltergeist to work with latest phantomjs (v2.0)
I was seeing the following error when running the tests with phantomjs v2.0:

    TypeError: undefined is not an object (evaluating 'phantom.args[0]')
    [...]/gems/poltergeist-1.3.0/lib/capybara/poltergeist/client/compiled/main.js:187 in global code

And the tests appeared to "hang" at this point.

The error appears to be related to an incompatibility between poltergeist
v1.3.0 and phantomjs v2.0 which has been fixed in this issue [1] and released
in poltergeist v1.6.0 [2].

The newer version of poltergeist seems to work ok with pre-2.0 versions of
phantomjs or at least it does with phantomjs v1.9.8.

The `Gemfile` seems to tie almost every gem to an exact version. Although I
don't understand why this is the case, I've followed the same convention for
now to avoid confusion.

[1]: https://github.com/teampoltergeist/poltergeist/issues/566
[2]: https://github.com/teampoltergeist/poltergeist/blob/master/CHANGELOG.md#160