Use slimmer 8.2.1
The new version adds an extra meta tag to the page. This will be
read by the analytics JavaScript if it is there (and give us the
ablity to segment performance metrics per rendering application
amongst other things)

It also uses the GOVUK_APP_NAME env variable rather than the
application class name so it can be cross referenced to other things in
the stack.

This contains alphagov/slimmer#128