Disable session store
As requested by @tekin as a precautionary measure for all frontend apps.
Apparently the Varnish cache strips session cookies on the way in and out
anyway, but at some point such cookies will selectively be allowed through
in order to implement authentication.

c.f. https://github.com/alphagov/government-frontend/pull/48

I added a spec to make it clear that this was done intentionally.