Jump to ruby 2.3.0
This also requires that we remove rubocop because the AST version is for
ruby 2.1.  We could also bump rubocop to get a version that includes an
AST for ruby 2.3, but in the next commit we were going to replace it with
the govuk-lin anyway.

Note that this commit results in a warning being generated in the tests:

    [warning] The response contained in an RestClient::Exception is now a RestClient::Response instead of a Net::HTTPResponse, please update your code

This is caused by gds-api-adapters when it converts an error response
into json for log purpopses.  It still actually does what we expect so
while it seems annoying, there's no harm in it (for now).  It appears
to be on the rest-client radar as https://github.com/rest-client/rest-client/issues/454
due to be fixed in rest-client 2.x (we're on latest 1.x).