Change which Mondays are included in the calculation.
The start date was considered the birth date/adoption date of the child. If the
child was born on a Monday, the child benefit started from the following Monday.

Under the new rules, we are going to assume the start date is the effective date
of the child benefit, i.e. the date the child benefit starts. If the effective
start date is a Monday, we also need to include that Monday in the calculation.

Updated the tests to account for the new "Monday" rules.