Refactor how the next Monday is calculated
The next monday is calculated in two places. Before the changes in
commit 994d1db, these calculations produced different result:

1. The adjusted_start_date attribute on the child is set only once and was
set to the Monday after the start date. This has since been changed to
also include the start date, if the start date is a Monday.

2. The code that determines how many weeks to calculate child benefit for
counts the number of Mondays in the tax year, including the first day of the
tax year if that happened to be a Monday.

Now both return the same date, so the next Monday calculation function (monday_on_or_after)
is being used in both instances.

The adjusted start date on the Starting Child is a property that changes
the child's start date to the next valid monday to comply with the rules.

Models shouldn't be aware of the rules of a calculator so this property has been removed.
The next Monday calculation function is instead being called  on the Starting Child
start date to extrapolate what the benefit start date would be.

The tests that previously tested the adjusted start date property have been removed and new
tests have been added to the child benefit tax calculator model.

monday_on_or_after uses the Rails DateAndTime helper functions to calculate the next Monday

See API documentation:
http://api.rubyonrails.org/classes/DateAndTime/Calculations.html#method-i-next_week