Only validate start date if is part year claim is set to yes
There's a bit of magic that happens when the page is loaded.
A starting child is created with empty start and end dates.
This empty child seems to exist behind the scenes, eventhough the user
can't see it.

Before the validation was restricted, If the user clicked calculate, then
the validation failed because the start date hadn't been entered, eventhough the
user cannot see the date fields on the page.

This seems to be occuring because the code in the starting child template loops through
the starting children on the model to know how many times to render the template.

This could be due to the fact that the children-template that renders
is part of the page is initially set to display:none, so at least one set of starting
child date fields are being created, but they are just not being showed to the user.

So when the user clicks "Calculate" these invisible fields are being validated eventhough
the user doesn't know they exist.