Allow a mix of full and part year children to be passed to the model
Pass in the start and end dates for tax year to the benefit amount calculation
for as many full year children that exist. Pass in the start and stop dates for the
part year children objects as before.

Only create children objects for the part year children so that their start and
stop dates can be reloaded in page when the page reloads (whether that's due to
the results or an error message being displayed)