Update advice about which dates to enter
The number of weeks included in the child benefit calculation has
changed. We are assuming the start date is the effective date of the
child benefit, i.e. the date the child benefit starts. If the effective
start date is a Monday, we are including it in the calculation.

The advice to users has been updated to reflect the changes.