Fix tax-year labels on mobile
This is the simplest fix that makes the page work.

A proper fix would be to refactor the CSS for this app, which has
some legacy issues, and doesn't play nicely with some of the common
styles from static.

That is a larger piece of work, that we should definately do in the
future, but for now, make it not broken for users.