Rename document type to “calculator”
We want to start using `document_type` in place of the deprecated `format` as the value of the `govuk:format` meta tag when using `govuk_components` (https://github.com/alphagov/static/pull/841).

In order to not change the current contents of the meta tag for calculators, this commit changes the `document_type` to match the current `format`.