Refactor the periods in drop downs
This commit makes changes to how and what the drop down box renders.

It introduces a date helper modules that contains methods that allow
for start and end year to be returned based on the arguments supplied.

This allow for more flexible dates for be derived and presented to the
user and reduce any previous confusion that may have existed.

In the past users were unable to select 2012 because the date range was
omitted as years past on.