Display new navigation for content tagged to taxons
As part of the rollout for the Parenting & Childcare taxonomy, we are
displaying new navigation elements on the Child Benefit Tax Calculator,
to provide a consistent UI for any users who are assigned to the “B”
bucket in the new Navigation A/B test.

This involves showing the taxon breadcrumbs, and the taxonomy sidebar.
On the landing page, the taxonomy sidebar replaces the existing sidebar.
On “main” pages, a taxonomy sidebar has been added, where there wasn’t
one before.

Note that naming has been kept to `EducationNavigation` and similar,
because that’s currently the naming in use by the cookie, and by
similar code in other applications. It has been kept like this to
maintain consistency and to assist future searches for code occurences.