Omit public_updated_at from payload
Publishing API will default to the time the new edition is published.

When creating the new content item for the form, this time was setting
`public_updated_at` but `first_published_at` was omitted – leading to
the Publishing API defaulting to publish time – which was 1 second
later than `public_updated_at`.

Removing `public_updated_at` from the payload avoids this discrepancy.

It also means `public_updated_at` is not updated every time there is a
deploy, as it does not change when the update is minor.