Fix and improve child benefit amount calculation
The way the calculator currently works out the child benefit within
the `benefits_claimed_amount` method is not easy to follow, and does not
handle a tax year with a rare scenario in that:
- it is a leap year
- it does not have one Monday in the last tax year month (April)

This means that for this current tax year (6 April 2019 to 5 April 2020)
it's calculating a result based on 53 weeks, when it should be 52 weeks.

The `benefits_claimed_amount` method has been refactored to use the
`Date.wday` method to find out how many Mondays are within a given date
range. This removes the previous complexity where it was being calculated
using an array that indexed the number of "weeks", and then looping through
them to count how many Mondays there are. At this point this is where it
can miscalculate by one Monday/week.

Some additional refactoring has also been done on the partial children
(`@starting_children`) calculation, where the start date and end date is
adjusted if they are empty or outside the tax year dates. This ensures
that only the applicable weeks/Mondays are included in the calculation.

Tests for tax year 2019/2020 have been added and current tests for the
benefit claim amount are hopefully clearer to follow for each tax year.