WIP: Remove years prior than 4 years ago
For question 2:
The options removed are years from 2012/2013 to 2015/2016, leaving only
options for years from 2016/2017 to 2020/2021.

For question 3:
The options removed are the years from 2011 to 2015, leaving only
options for years from 2016 to 2021.

When removing those options, I also refactored the code so that we don't
have to remove or add options in the future. Now, the provided options
are always:
- for question 2: the current year + the four previous years
- for question 3: the current year + the four previous years + next year

Many tests were dependant on the actual current date. This means we had
to edit them every year. I deleted some of the ones related to old years
we no longer support and changed others to be independent of the current
date using Timecop.