Reintroduce previous years
We initially removed support for years before 2016 in #853, including
removing tests that applied only to these years (and moving some other
tests to later years). However, HMRC later realized that there are some
cases where years before 2016 are still needed, so we were asked to
restore them.

This commit therefore restores support for years 2012–16, including
tests that apply to those years, but tries to retain some of the
refactoring we'd done in #853 so that we can have the benefit of that
still.