Switch asset path and use ASSET_HOST env var for asset host
This is in preparation for these assets to be hosted on www.gov.uk,
it removes the GOVUK_ASSET_HOST env var that is set globally and instead
allows setting the asset host with an ASSET_HOST environment variable.

The expectation is that this environment variable won't be set in
production environments and instead only used in dev/test environments
for working around a lack of proxying from the router host.