Limit potential tax years to those with defined rates
A situation can occur where the option for a particular tax year is
displayed even before the rates have been defined for it. It should be
limited to within this range. (This likely occurs in a new year, because
we decide the years to show based on the current calendar year, not the
current tax year, but rates are based on tax years).

This was exposed by #1023, which altered the validation — but it is
likely that the radio option for a tax year was incorrectly displayed
even before that.